Renaming and some refinement
