+++ perfomance-tests/shell/run-tests-sequential.sh
