+++ perfomance-tests/tests/b2b-mp/sapi/tests/product-page.js
@@ -7,7 +7,7 @@ options.scenarios = {
         exec: 'executeProductPageScenario',
         executor: 'shared-iterations',
         vus: 1,
-        iterations: 2,
+        iterations: 10,
     },
 };
