+++ perfomance-tests/tests/b2b-mp/storefront/tests/B2B-MP-S2-checkout-flow-1-item.js
@@ -3,15 +3,17 @@ import { loadDefaultOptions } from "../../../../lib/utils.js";
 
 export const options = loadDefaultOptions();
 options.scenarios = {
-    Checkout_70_items: {
+    S2_Checkout: {
         exec: 'executeCheckoutScenario',
         executor: 'shared-iterations',
         env: {
-            numberOfItems: __ENV.numberOfItems || '70'
+            numberOfItems: __ENV.numberOfItems || '1'
         }
     },
 };
 
+//scenario objects must be created outside any function used in execute phase since some initialization actions are done on
+//K6 "init" stage (in the current implementation such init action are done in class constructor).
 const checkoutScenario = new CheckoutScenario();
 
 export function executeCheckoutScenario() {
